optimize requestListener
